test: Splite GeoDDCoordinates tests in multiple files
