+++ tests/PowerUtils.Geolocation.Tests/GeoDDCoordinates/HashCodeTests.cs
@@ -6,6 +6,38 @@ namespace PowerUtils.Geolocation.Tests.GeoDDCoordinates
     public sealed class HashCodeTests
     {
         [Fact]
+        public void EqualsProperties_ComparisonHashCodes_True()
+        {
+            // Arrange
+            var left = new GeoDDCoordinate(1.54, 54.1272);
+            var right = new GeoDDCoordinate(1.54, 54.1272);
+
+
+            // Act
+            var act = left.GetHashCode() == right.GetHashCode();
+
+
+            // Assert
+            act.Should().BeTrue();
+        }
+
+        [Fact]
+        public void DifferentsProperties_ComparisonHashCodes_False()
+        {
+            // Arrange
+            var left = new GeoDDCoordinate(1.54, 5.1272);
+            var right = new GeoDDCoordinate(-1.54, 54.1272);
+
+
+            // Act
+            var act = left.GetHashCode() == right.GetHashCode();
+
+
+            // Assert
+            act.Should().BeFalse();
+        }
+
+        [Fact]
         public void HashCode_ArithmeticMutant_MultiplicationVsDivision()
         {
             // This test targets the Math.Round(value / HASH_TOLERANCE) * HASH_TOLERANCE formula
@@ -245,6 +277,100 @@ namespace PowerUtils.Geolocation.Tests.GeoDDCoordinates
             // Should not be extreme values that might indicate overflow
             hashCode.Should().BeInRange(int.MinValue + 1000, int.MaxValue - 1000,
                 "hash should not be near overflow boundaries");
+        }
+
+        [Fact]
+        public void NearlyIdenticalCoordinates_GetHashCode_ShouldBeConsistentWithEquality()
+        {
+            // Arrange
+            var coord1 = new GeoDDCoordinate(45.123456789012, -90.987654321098);
+            var coord2 = new GeoDDCoordinate(45.123456789012, -90.987654321098);
+            var coord3 = new GeoDDCoordinate(45.123456789013, -90.987654321099); // Tiny difference
+
+
+            // Act
+            var hash1 = coord1.GetHashCode();
+            var hash2 = coord2.GetHashCode();
+            var hash3 = coord3.GetHashCode();
+
+            var equals_1_2 = coord1 == coord2;
+            var equals_1_3 = coord1 == coord3;
+
+
+            // Assert
+            equals_1_2.Should().BeTrue("Identical coordinates should be equal");
+            hash1.Should().Be(hash2, "Equal objects should have equal hash codes");
+
+            if(equals_1_3)
+            {
+                hash1.Should().Be(hash3, "If objects are equal, hash codes must be equal");
+            }
+            // Note: Different objects can have the same hash code, but equal objects must have the same hash code
+        }
+
+        [Fact]
+        public void HashCodeContractValidation_EqualCoordinatesWithinTolerance_ShouldHaveSameHashCode()
+        {
+            // Arrange - Create coordinates that are equal within tolerance (1e-12)
+            var base1 = 45.123456789012;
+            var base2 = -90.987654321098;
+
+            // Create a coordinate with differences exactly at the tolerance boundary
+            var diff = 1e-13; // Just within tolerance
+            var coord1 = new GeoDDCoordinate(base1, base2);
+            var coord2 = new GeoDDCoordinate(base1 + diff, base2 + diff);
+
+            // Act
+            var areEqual = coord1 == coord2;
+            var hash1 = coord1.GetHashCode();
+            var hash2 = coord2.GetHashCode();
+
+            // Assert - This is the critical hash code contract test
+            if(areEqual)
+            {
+                hash1.Should().Be(hash2,
+                    "Hash code contract violation: equal objects must have equal hash codes. " +
+                    $"Tolerance: 1e-12, Difference: {diff}, Equal: {areEqual}");
+            }
+
+            // Verify they are indeed equal within tolerance
+            areEqual.Should().BeTrue("Coordinates within tolerance should be equal");
+        }
+
+        [Fact]
+        public void NearlyIdenticalCoordinates_InHashSet_ShouldBehaveConsistently()
+        {
+            // Collection Behavior Tests
+
+            // Arrange
+            var hashSet = new System.Collections.Generic.HashSet<GeoDDCoordinate>();
+
+            var coord1 = new GeoDDCoordinate(45.123456789012, -90.987654321098);
+            var coord2 = new GeoDDCoordinate(45.123456789012, -90.987654321098);
+            var coord3 = new GeoDDCoordinate(45.123456789013, -90.987654321099); // Tiny difference
+
+
+            // Act
+            hashSet.Add(coord1);
+            var added2 = hashSet.Add(coord2); // Should not be added if equal to coord1
+            var added3 = hashSet.Add(coord3); // May or may not be added depending on equality
+
+
+            // Assert
+            added2.Should().BeFalse("Identical coordinate should not be added again");
+            hashSet.Should().Contain(coord1);
+            hashSet.Should().Contain(coord2); // Should find it even if not added
+
+            // Document behavior with nearly identical coordinates
+            if(coord1 == coord3)
+            {
+                added3.Should().BeFalse("Nearly identical coordinates should not be added if considered equal");
+                hashSet.Should().Contain(coord3);
+            }
+            else
+            {
+                added3.Should().BeTrue("Different coordinates should be added");
+            }
         }
     }
 }
